% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gengo2seireki.R
\name{gengo2seireki}
\alias{gengo2seireki}
\title{Translate year in gengo into year in seireki}
\usage{
gengo2seireki(x)
}
\arguments{
\item{x}{string which indicates date.}
}
\description{
Translate year in gengo into year in seireki. Meiji, Taisyo, Syowa, Heisei, and Reiwa can be recognized.
}
\examples{
gengo2seireki("H31.4.30")
}
