% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jymd.R
\name{jymd}
\alias{jymd}
\alias{jymd_h}
\alias{jymd_hm}
\alias{jymd_hms}
\title{Parse dates with \strong{y}ear, \strong{mo}nth, \strong{d}ay, and Japanese nengo components}
\usage{
jymd(..., quiet = FALSE, tz = "Asia/Tokyo",
  locale = Sys.getlocale("LC_TIME"), truncated = 0)

jymd_h(..., quiet = FALSE, tz = "Asia/Tokyo",
  locale = Sys.getlocale("LC_TIME"), truncated = 0)

jymd_hm(..., quiet = FALSE, tz = "Asia/Tokyo",
  locale = Sys.getlocale("LC_TIME"), truncated = 0)

jymd_hms(..., quiet = FALSE, tz = "Asia/Tokyo",
  locale = Sys.getlocale("LC_TIME"), truncated = 0)
}
\arguments{
\item{...}{a character or numeric vector of suspected dates}

\item{quiet}{logical. When TRUE function evalueates without displaying
customary messages.}

\item{tz}{Time zone indicator. If \code{NULL} a Date object is returned. Otherwise a POSIXct with time zone attribute set to tz.}

\item{locale}{locale to be used, see \link{locales}. On linux systems you
can use \code{system("locale -a")} to list all the installed locales.}

\item{truncated}{integer. Number of formats that can be truncated.}
}
\description{
An wrapper of \code{\link[lubridate:ymd]{lubridate::ymd()}} to enable paring Japanese gengo. Meiji, Taisyo, Syowa, Heisei, and Reiwa can be recognized.
}
\note{
\code{tz} argument has default value as \code{"Asia/Tokyo"} unlike \code{NULL} in \code{\link[lubridate:ymd]{lubridate::ymd()}}.
}
